cached: Remove unnecessary overwrite of handler
Previously, when completing an entry, we would set the xcache handler of
the peer request's cio as NoEntry. This was not needed due to the fact
that when we get a new peer request, we set this handler immediately to
the handler returned to us by lookup/insert.

Also, this leads to severe problems if the peer request we are
completing is the ce's itself (e.g. when doing a flush). In this case,
the following request that gets this ce will not be able to flush it.