xcache: Add O(1) LRU implementation
This implementation utilizes the hash table to create an indexed doubly
linked list. The list cn be visualised as so:

<-- Most recently used            Least recently used -->
MRU <-> MRU+ <-> MRU++ <-> ... <-> LRU^^ <-> LRU^ <-> LRU

Retrieving the LRU element is fairly easy; it's always the last one.

Updating an element that is in the middle of the chain is also easy.
Since we have already used the hash table to find the cache entry, we
update the pointers of this element and append it at the top (MRU) of
t at the top (MRU) of the chain.