xcache: Explicitly set flag to use rm tables
Introduce a new flag (XCACHE_USE_RMTABLE) that will be set only by the
peers that want to use the rm_table. The rest of the peers don't need to
set anything. This should boost speed for the latter peers, since they
won't unnecessarily try to acquire the rm_lock on inserts and puts

Changes:

* The "post_evict" hook is no longer necessary and is superseded by this
  flag.
* This flag is passed during xcache_init in the "flags" parameter.

Conflicts:
	xseg/xtypes/xcache.h