cached: Manually evict objects to regain buckets
If the bucket pool is empty but no object is in the process of
eviction, we have to manually evict the LRU object ourselves to gain
free space.

We do so after we have finished with dispatching a request and before
entering the generic peerd loop, and only when the following
requirements are met:

1. The bucket pool is empty
2. There are requests waiting at the bucket_waitq
3. There are no evicted objects

The functionality for the first two requirements has are already been
implemented. To achieve the third requirement, this commit introduces
a simple track of evicted objects.