cached: Fix deferred_workq loop issue
This commit solves the problem of having to defer a work whilst being
in the deferred_workq, which causes the thread to enter a never-ending
loop. The problem description follows:

~~~
The deferred_workq is used to enqueue jobs that have to send requests to
the blocker but can't, because another job is already doing so.
Consecutively, this workq is signaled when a job has finished
issuing requests to the blocker.

However, if the deferred_workq is being signaled and a work in that
queue cannot be executed immediately, it must be enqueued once more in
this queue, which would ultimately lead to it being signaled again, then
enqueued again, then signaled again etc...
~~~

The solution is the utilisation of a waitq as the middle layer between
deferring the work and enqueueing it to the ce's workq.

Simply put, when we have to defer a workq, we enqueue it in a
pending_waitq. The signaling condition of this waitq must be the same as
the execution condition of the job i.e. the condition that allows a job
to be executed must be the same as the condition that can signal the
waitq. Finally, in this middle layer, when the work can be signaled, it
will attempt to enqueue itself in the ce's workq, thus running safely
with the ce->lock.