tests: Fix locking test for cached
In cached tests, the role of blocker is taken by cached and the blocker
(filed) is called "realblocker" instead. In "test_locking" however,
blocker (filed) must be restarted during the test, and this switch
results in restarting cached instead of filed.

Amend this behavior by creating an overwritable function that restarts
peers and can be controlled in class CachedTest.