cached: Remove unneeded warning
Since buckets are still in experimental phase, we have employed the
concept of safety switch in the function that frees them.

The safety switch's purpose is to warn us when we try to free a bucket
range that has buckets with status other than valid. There are places in
the code where this is acceptable and places where it's not, so the
switch must be set accordingly.

In our case, after we have flushed dirty buckets and want to instantly
free them (due to lack of space), we would falsely set the switch as on.
Yet, it is acceptable to flush the dirty buckets, receive a write in the
meantime and then free all but the dirty buckets.

For this reason, this switch is now turned off for the above case.