cached: Add termination condition for cached
The usual termination condition for cached was not sufficient since,
while there may not be any requests pending, there could be dirty
data or flushes happening in the background.

For this reason, we also check if all buckets are either valid or
invalid. This implicitly means that there will be no dirty data when
we exit and no background flushes.

This commit fixes Bug #3783