cached: Add support for info requests
When an info request is accepted, we check if the object exists in
cache and keep its cache handler. Then, we forward the request to the
blocker. When the blocker answers, we either propagate the blocker's
answer to the asker (if we're on write-through mode or haven't cached the
object) or check if we have written to the object beyond the blocker and
respond accordingly.