xcache: Fix behavior on close and invalidate
xcache_close uses only put to release its reference to the entry. The
peer that uses xcache must use xcache_close *only* when it has released
any references *it* has to the entries.

Moreover, xcache_invalidate now updates the state of the entry, which
for the time being is useful only to xcache_put.