xcache: Fix the scenario when we evict all entries
Since there are cases where all entries cannot be evicted
simultaneously, we need a way to defer their eviction.

This commit introduces the following procedure. Check if the entry can
be inserted in the rm_entries hash table and if not, return and insert
that entry again in the LRU queue.

The only drawback is that the LRU entry becomes the MRU entry, but this
scenario should happen only when xcache is finalizing, and thus should
not bother us.