glusterd: Create initial working implementation
Modify the skeleton of sosd to the GlusterFS needs. We have substituted
the RADOS pool for the GlusterFS volume and the RADOS objects for the
files of GlusterFS. The rest of the logic remains intact.

The current implementation only supports read/write operations. Besides
working on the support for the rest of the operations (X_HASH, X_COPY...),
we must take into account the following caveats:

* RADOS can read/write to objects without the need to "open" or "create"
  them. This is not the case for GlusterFS, which uses filesystem
  semantics. Unfortunately, the open/create operations are blocking,
  which means that we have to find a way around that.
* The `stat`,`unlink` operations are also blocking.
* There does not seem to be a way to create a callback for notifications
  on the file (like `rados_watch`)

Also, we need to add more comments and adhere to a sensible coding
style.