glusterd: Add support for more XSEG operations
Support the following operations:

* X_COPY
* X_HASH
* X_LOCK
* X_UNLOCK

(Note that currently the X_UNLOCK has the semantics of breaking the lock.
This will be changed in the following commit messages)

The operations that are unsupported are:

* X_DELETE
* X_INFO

Also, we try to reuse as much as possible the same functions to make the
code more readable and easier to debug. This is a necessity since e.g.
in X_HASH, we need to read/write/open many times different files and
handle properly the half-reads and half-writes. Reusing the code
simplifies these subtasks and reduces the problem to just passing at the
appropriate read/write/open functions the correct targets and buffers.