tests: Add tests for glusterd
The tests that glusterd is tested against are the common tests for all
blocker peers.

Note that for the tests, the libgfapi-python library is required, in
order to cleanup the volume before every test run.

WARNING: Due to instability issues of libgfapi-python, if the whole
suite is instructed to run, it will probably result in a segfault of
libgfapi. Thus, the user is suggested to run each test individually.