archipelago: Add gluster configuration example
This example configuration replaces the default blocker (filed) with
glusterd. By default, it expects that the Archipelago host is part of
the Gluster pool and that the volume is called "gv0". Moreover, glusterd
is set by default in synchronous mode (async = 0).

This configuration can be enabled with the following command:

cp /etc/archipelago/archipelago.conf.gluster_example \
	/etc/archipelago/archipelago.conf